fix windows move and resize
